refactoring bin script [skip ci] [skip codecov]
